+++ src/main/java/com/estsoft/findmember_team01/admin/controller/AdminController.java
@@ -3,13 +3,13 @@ package com.estsoft.findmember_team01.admin.controller;
 import com.estsoft.findmember_team01.admin.dto.MemberViewResponse;
 import com.estsoft.findmember_team01.admin.dto.UpdateLevelRequest;
 import com.estsoft.findmember_team01.admin.service.AdminService;
-import com.estsoft.findmember_team01.application.domain.Recruitment;
-import com.estsoft.findmember_team01.application.dto.RecruitmentRequest;
-import com.estsoft.findmember_team01.application.dto.RecruitmentResponse;
-import com.estsoft.findmember_team01.application.dto.RecruitmentViewResponse;
-import com.estsoft.findmember_team01.application.service.RecruitmentService;
 import com.estsoft.findmember_team01.member.domain.Member;
 import com.estsoft.findmember_team01.member.service.MemberService;
+import com.estsoft.findmember_team01.recruitment.domain.Recruitment;
+import com.estsoft.findmember_team01.recruitment.dto.RecruitmentRequest;
+import com.estsoft.findmember_team01.recruitment.dto.RecruitmentResponse;
+import com.estsoft.findmember_team01.recruitment.dto.RecruitmentViewResponse;
+import com.estsoft.findmember_team01.recruitment.service.RecruitmentService;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
