+++ src/main/java/com/estsoft/findmember_team01/comment/repository/CommentRepository.java
@@ -1,10 +1,9 @@
-package com.estsoft.findmember_team01.information.repository;
+package com.estsoft.findmember_team01.comment.repository;
 
-import com.estsoft.findmember_team01.information.domain.Comment;
+import com.estsoft.findmember_team01.comment.domain.Comment;
+import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
 
 @Repository
 public interface CommentRepository extends JpaRepository<Comment, Long> {
