+++ src/main/java/com/estsoft/findmember_team01/login/handler/CustomLogoutSuccessHandler.java
@@ -1,4 +1,4 @@
-package com.estsoft.findmember_team01.member.handler;
+package com.estsoft.findmember_team01.login.handler;
 
 import com.estsoft.findmember_team01.member.repository.MemberRepository;
 import jakarta.servlet.http.HttpServletRequest;
