+++ src/main/java/com/estsoft/findmember_team01/member/controller/ExpController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-public class RoleController {
+public class ExpController {
 
     private final MemberService memberService;
